Replaced float by double in atom::ad_covalent_radii and scoring_function::vdw

+++ src/atom.cpp
@@ -40,7 +40,7 @@ const array<string, atom::n> atom::ad_strings =
 };
 
 //! Covalent radii of AutoDock4 atom types, factorized by 1.1 for extra allowance. http://en.wikipedia.org/wiki/Atomic_radii_of_the_elements_(data_page)
-const array<float, atom::n> atom::ad_covalent_radii =
+const array<double, atom::n> atom::ad_covalent_radii =
 {
 	0.407, //  0 = H , 0.407 = 1.1 * 0.37
 	0.407, //  1 = HD, 0.407 = 1.1 * 0.37
@@ -115,7 +115,7 @@ const array<size_t, atom::n> atom::ad_to_xs =
 atom::atom(const string& line) :
 	serial(stoul(line.substr(6, 5))),
 	name(line.substr(12, 4)),
-	coordinate({stof(line.substr(30, 8)), stof(line.substr(38, 8)), stof(line.substr(46, 8))}),
+	coordinate({stod(line.substr(30, 8)), stod(line.substr(38, 8)), stod(line.substr(46, 8))}),
 	ad(find(ad_strings.cbegin(), ad_strings.cend(), line.substr(77, isspace(line[78]) ? 1 : 2)) - ad_strings.cbegin()),
 	xs(ad_to_xs[ad])
 {
